Fix some warnings in examples

+++ examples/FeatureShowcase/Source/PrimitiveBatchTest/PrimitiveBatchTest.cpp
@@ -50,8 +50,8 @@ void PrimitiveBatchTest::Draw()
     commandList->SetRenderPass(std::move(pass));
 
     auto projectionMatrix = Matrix4x4::CreateOrthographicLH(
-        presentationParameters.BackBufferWidth,
-        presentationParameters.BackBufferHeight,
+        static_cast<float>(presentationParameters.BackBufferWidth),
+        static_cast<float>(presentationParameters.BackBufferHeight),
         0.0f,
         100.0f);
